pyproject.toml
[project]
name = "hailo-llm-deploy"
version = "0.1.0"
description = "HuggingFace LLM fine-tuning and Hailo NPU edge deployment pipeline"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "torch>=2.0",
    "transformers>=4.40",
    "peft>=0.10",
    "trl>=0.8",
    "datasets>=2.0",
]

[project.optional-dependencies]
eval = ["rouge-score", "bert-score", "numpy"]
export = ["optimum", "onnx", "onnxruntime"]
hailo = ["numpy"]  # hailo_sdk_client must be installed via Hailo AI SW Suite
dev = ["ruff", "pytest"]

[project.scripts]
hailo-llm-deploy = "hailo_llm_deploy.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["hailo_llm_deploy*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
